refactor: BoardStateEvaluator -> BoardStateUtils + ajuste chamadas para evitar passagem de Game como parâmetro
