+++ backend/src/main/java/com/chess/service/game/GameService.java
@@ -1,5 +1,6 @@
 package com.chess.service.game;
 
+import java.util.List;
 import java.util.Objects;
 import java.util.Timer;
 import java.util.TimerTask;
@@ -16,6 +17,7 @@ import com.chess.entity.move.Move;
 import com.chess.entity.piece.Piece;
 import com.chess.service.move.MoveExecutor;
 import com.chess.service.move.MoveValidator;
+import com.chess.utils.BoardStateUtils;
 
 public class GameService {
     
@@ -93,10 +95,12 @@ public class GameService {
 
         MoveValidator validator = MoveValidator.of(game.getCurrentBoard());
         Move move = validator.createMove(from, to, promotionPiece);
+
         BoardBuilder nextBoardBuilder = moveExecutor.executeMove(game.getCurrentBoard(), move);
+        List<Board> boardHistory = game.getBoardHistory();
         
         // Validar estado do BoardBuilder e instanciar Board
-        nextBoardBuilder.setBoardState(BoardStateEvaluator.evaluateState(nextBoardBuilder, game));
+        nextBoardBuilder.setBoardState(BoardStateUtils.evaluateState(nextBoardBuilder, boardHistory));
         Board nextBoard = nextBoardBuilder.build();
 
         game.commitMove(move, nextBoard);
@@ -179,7 +183,7 @@ public class GameService {
         Color opponent = playerWhoRanOutOfTime.opposite();
 
         // Verifico se o VENCEDOR (por tempo) tem material suficiente para vencer
-        boolean opponentHasMatingMaterial = BoardStateEvaluator.hasMatingMaterial(getCurrentBoard(), opponent);
+        boolean opponentHasMatingMaterial = BoardStateUtils.hasMatingMaterial(getCurrentBoard(), opponent);
 
         if (opponentHasMatingMaterial) {
             game.timeoutLoss(playerWhoRanOutOfTime);
